+++ packages/core/src/modules/openid/features/OpenIDProofPresentation/OpenIDProofRequestDisplay.tsx
@@ -1,3 +1,4 @@
+import React from 'react'
 import { StyleSheet, View, ScrollView } from 'react-native'
 
 import { FormattedSubmission, OpenId4VPRequestRecord } from '../../types'